+++ Source/Scene/PointCloud3DTileContent.js
@@ -16,6 +16,7 @@ define([
         '../Core/Matrix3',
         '../Core/Matrix4',
         '../Core/oneTimeWarning',
+        '../Core/OrthographicFrustum',
         '../Core/Plane',
         '../Core/PrimitiveType',
         '../Core/RuntimeError',
@@ -53,6 +54,7 @@ define([
         Matrix3,
         Matrix4,
         oneTimeWarning,
+        OrthographicFrustum,
         Plane,
         PrimitiveType,
         RuntimeError,
@@ -151,6 +153,13 @@ define([
          * @inheritdoc Cesium3DTileContent#featurePropertiesDirty
          */
         this.featurePropertiesDirty = false;
+
+        // Options for geometric error based attenuation
+        this._attenuation = false;
+        this._geometricErrorScale = undefined;
+        this._maximumAttenuation = undefined;
+        this._baseResolution = undefined;
+        this._baseResolutionApproximation = undefined;
 
         initialize(this, arrayBuffer, byteOffset);
     }
@@ -450,9 +459,17 @@ define([
         content._hasColors = defined(colors);
         content._hasNormals = defined(normals);
         content._hasBatchIds = defined(batchIds);
+
+        // Compute an approximation for base resolution in case it isn't given.
+        // Assume a uniform distribution of points in cubical cells throughout the
+        // bounding sphere around the tile.
+        // Typical use case is leaves, where lower estimates of interpoint distance might
+        // lead to underattenuation.
+        var sphereVolume = content._tile.contentBoundingVolume.boundingSphere.volume();
+        content._baseResolutionApproximation = Math.cbrt(sphereVolume / pointsLength);
     }
 
-    var scratchPointSizeAndTilesetTime = new Cartesian2();
+    var scratchPointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier = new Cartesian4();
 
     var positionLocation = 0;
     var colorLocation = 1;
@@ -524,10 +541,30 @@ define([
         }
 
         var uniformMap = {
-            u_pointSizeAndTilesetTime : function() {
-                scratchPointSizeAndTilesetTime.x = content._pointSize;
-                scratchPointSizeAndTilesetTime.y = content._tileset.timeSinceLoad;
-                return scratchPointSizeAndTilesetTime;
+            u_pointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier : function() {
+                var scratch = scratchPointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier;
+                scratch.x = content._attenuation ? content._maximumAttenuation : content._pointSize;
+                scratch.y = content._tileset.timeSinceLoad;
+
+                if (content._attenuation) {
+                    var geometricError = content.tile.geometricError;
+                    if (geometricError === 0) {
+                        geometricError = defined(content._baseResolution) ? content._baseResolution : content._baseResolutionApproximation;
+                    }
+                    var frustum = frameState.camera.frustum;
+                    var depthMultiplier;
+                    // Attenuation is maximumAttenuation in 2D/ortho
+                    if (frameState.mode === SceneMode.SCENE2D || frustum instanceof OrthographicFrustum) {
+                        depthMultiplier = Number.POSITIVE_INFINITY;
+                    } else {
+                        depthMultiplier = context.drawingBufferHeight / frameState.camera.frustum.sseDenominator;
+                    }
+
+                    scratch.z = geometricError * content._geometricErrorScale;
+                    scratch.w = depthMultiplier;
+                }
+
+                return scratch;
             },
             u_highlightColor : function() {
                 return content._highlightColor;
@@ -818,6 +855,7 @@ define([
         var backFaceCulling = content._backFaceCulling;
         var vertexArray = content._drawCommand.vertexArray;
         var clippingPlanes = content._tileset.clippingPlanes;
+        var attenuation = content._attenuation;
 
         var colorStyleFunction;
         var showStyleFunction;
@@ -930,11 +968,16 @@ define([
 
         var vs = 'attribute vec3 a_position; \n' +
                  'varying vec4 v_color; \n' +
-                 'uniform vec2 u_pointSizeAndTilesetTime; \n' +
+                 'uniform vec4 u_pointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier; \n' +
                  'uniform vec4 u_constantColor; \n' +
-                 'uniform vec4 u_highlightColor; \n' +
-                 'float u_pointSize; \n' +
-                 'float u_tilesetTime; \n';
+                 'uniform vec4 u_highlightColor; \n';
+        vs += 'float u_pointSize; \n' +
+              'float u_tilesetTime; \n';
+
+        if (attenuation) {
+            vs += 'float u_geometricError; \n' +
+                  'float u_depthMultiplier; \n';
+        }
 
         vs += attributeDeclarations;
 
@@ -983,8 +1026,13 @@ define([
 
         vs += 'void main() \n' +
               '{ \n' +
-              '    u_pointSize = u_pointSizeAndTilesetTime.x; \n' +
-              '    u_tilesetTime = u_pointSizeAndTilesetTime.y; \n';
+              '    u_pointSize = u_pointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier.x; \n' +
+              '    u_tilesetTime = u_pointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier.y; \n';
+
+        if (attenuation) {
+            vs += '    u_geometricError = u_pointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier.z; \n' +
+                  '    u_depthMultiplier = u_pointSizeAndTilesetTimeAndGeometricErrorAndDepthMultiplier.w; \n';
+        }
 
         if (usesColors) {
             if (isTranslucent) {
@@ -1032,6 +1080,11 @@ define([
 
         if (hasPointSizeStyle) {
             vs += '    gl_PointSize = getPointSizeFromStyle(position, position_absolute, color, normal); \n';
+        } else if (attenuation) {
+            vs += '    vec4 positionEC = czm_modelView * vec4(position, 1.0); \n' +
+                  '    float depth = -positionEC.z; \n' +
+                  // compute SSE for this point
+                  '    gl_PointSize = min((u_geometricError / depth) * u_depthMultiplier, u_pointSize); \n';
         } else {
             vs += '    gl_PointSize = u_pointSize; \n';
         }
@@ -1042,7 +1095,7 @@ define([
             vs += '    normal = czm_normal * normal; \n' +
                   '    float diffuseStrength = czm_getLambertDiffuse(czm_sunDirectionEC, normal); \n' +
                   '    diffuseStrength = max(diffuseStrength, 0.4); \n' + // Apply some ambient lighting
-                  '    color *= diffuseStrength; \n';
+                  '    color.xyz *= diffuseStrength; \n';
         }
 
         vs += '    v_color = color; \n' +
@@ -1254,6 +1307,19 @@ define([
             this._unionClippingRegions = unionClippingRegions;
 
             createShaders(this, frameState, tileset.style);
+        }
+
+        // Update attenuation
+        var pointCloudShading = this._tileset.pointCloudShading;
+        if (defined(pointCloudShading)) {
+            var formerAttenuation = this._attenuation;
+            this._attenuation = pointCloudShading.attenuation;
+            this._geometricErrorScale = pointCloudShading.geometricErrorScale;
+            this._maximumAttenuation = defined(pointCloudShading.maximumAttenuation) ? pointCloudShading.maximumAttenuation : tileset.maximumScreenSpaceError;
+            this._baseResolution = pointCloudShading.baseResolution;
+            if (this._attenuation !== formerAttenuation) {
+                createShaders(this, frameState, tileset.style);
+            }
         }
 
         if (updateModelMatrix) {