+++ Specs/Widgets/Cesium3DTilesInspector/Cesium3DTilesInspectorViewModelSpec.js
@@ -161,6 +161,55 @@ defineSuite([
             viewModel.showUrl = false;
             expect(viewModel.tileset.debugShowUrl).toBe(false);
         });
+
+        it('pointCloudShading', function() {
+            viewModel.pointCloudShading = true;
+            expect(viewModel.tileset.pointCloudShading.attenuation).toBe(true);
+            viewModel.pointCloudShading = false;
+            expect(viewModel.tileset.pointCloudShading.attenuation).toBe(false);
+        });
+
+        it('geometricErrorScale', function() {
+            viewModel.geometricErrorScale = 1.0;
+            expect(viewModel.tileset.pointCloudShading.geometricErrorScale).toBe(1.0);
+            viewModel.geometricErrorScale = 0.0;
+            expect(viewModel.tileset.pointCloudShading.geometricErrorScale).toBe(0.0);
+        });
+
+        it('maximumAttenuation', function() {
+            viewModel.maximumAttenuation = 1.0;
+            expect(viewModel.tileset.pointCloudShading.maximumAttenuation).toBe(1.0);
+            viewModel.maximumAttenuation = 0.0;
+            expect(viewModel.tileset.pointCloudShading.maximumAttenuation).not.toBeDefined();
+        });
+
+        it('baseResolution', function() {
+            viewModel.baseResolution = 1.0;
+            expect(viewModel.tileset.pointCloudShading.baseResolution).toBe(1.0);
+            viewModel.baseResolution = 0.0;
+            expect(viewModel.tileset.pointCloudShading.baseResolution).not.toBeDefined();
+        });
+
+        it('eyeDomeLighting', function() {
+            viewModel.eyeDomeLighting = true;
+            expect(viewModel.tileset.pointCloudShading.eyeDomeLighting).toBe(true);
+            viewModel.eyeDomeLighting = false;
+            expect(viewModel.tileset.pointCloudShading.eyeDomeLighting).toBe(false);
+        });
+
+        it('eyeDomeLightingStrength', function() {
+            viewModel.eyeDomeLightingStrength = 1.0;
+            expect(viewModel.tileset.pointCloudShading.eyeDomeLightingStrength).toBe(1.0);
+            viewModel.eyeDomeLightingStrength = 0.0;
+            expect(viewModel.tileset.pointCloudShading.eyeDomeLightingStrength).toBe(0.0);
+        });
+
+        it('eyeDomeLightingRadius', function() {
+            viewModel.eyeDomeLightingRadius = 1.0;
+            expect(viewModel.tileset.pointCloudShading.eyeDomeLightingRadius).toBe(1.0);
+            viewModel.eyeDomeLightingRadius = 0.0;
+            expect(viewModel.tileset.pointCloudShading.eyeDomeLightingRadius).toBe(0.0);
+        });
     });
 
     describe('update options', function() {